requirements.txt
openai==1.3.0
Django==3.2.10
requests
gTTS
Pillow
pydub==0.25.1
moviepy==1.0.3
python-dotenv